+++ changelog.md
@@ -8,7 +8,7 @@
 
 ## [5.3.1] - 2025-08-08
 
-[commit](https://github.com/sbrstrkkdwmdr/ssob)</br>
+[commit](https://github.com/sbrstrkkdwmdr/ssob/commit/f492b83b50570e9cb43b4cf1d74f2b6849581a99)</br>
 
 ### Fixed