+++ changelog.md
@@ -2,7 +2,12 @@
 
 # changelog
 
+
 ## [WIP] - xxxx-xx-xx
+
+[commit](https://github.com/sbrstrkkdwmdr/ssob)</br>
+
+## [5.3.2] - 2025-08-26
 
 [commit](https://github.com/sbrstrkkdwmdr/ssob)</br>
 
@@ -11,6 +16,7 @@
 -   score list sending multiple errors if empty
 -   "F if pass" on score list commands
 -   missing prefix in command examples
+-   commands not accepting mapid via flags
 
 ## [5.3.1] - 2025-08-08